requirements.txt
pandas>=1.5.0
numpy>=1.23.0
yfinance>=0.2.0
fredapi>=0.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
pyyaml>=6.0
scipy>=1.9.0
pytest>=7.0.0
black>=22.0.0